ml-services/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
joblib
pymongo
python-dotenv
